third_party/WebKit/LayoutTests/fast/js/exception-properties-expected.txt

Test for correct properties on Error objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS enumerableProperties(error) is []
FAIL enumerableProperties(nativeError) should be stack,line,sourceURL. Was .
PASS Object.getPrototypeOf(nativeError).name is "RangeError"
PASS Object.getPrototypeOf(nativeError).message is ""
PASS successfullyParsed is true

TEST COMPLETE
